fix: language configurations are now correctly included in the build
The fault was that they weren't copied to `dist/` with the built files, but referenced locally.
This meant they failed to occur in the published version of the extension.
Update syntaxes project `build` target with verbose commands syntax to copy them over correctly.
Retarget reference to distribution folder, so local builds won't present a false positive anymore.
This change correctly includes them in the bundle.
The required auto-closing/indentation/comment declarations are now contributed to ccini! ^_^.